+++ NEWS.md
@@ -6,3 +6,5 @@ how to get started.
 **Breaking Changes to dev version**
 
 * `with_touchstone_lib()` was renamed to `run_script()`.
+* `benchmark_run_ref()` does not accept a character input for code expressions
+  anymore (#62).